+++ src/ui/views/Gallery/components/Card/CardMedia.svelte
@@ -1,3 +1,3 @@
-<div class="projects--gallery--card__media" on:keypress on:click>
+<div class="projects--gallery--card__media" on:keypress on:click on:touchstart on:touchmove on:touchend>
   <slot />
 </div>